+++ ui/src/app/core/accessories/accessory-info/accessory-info.component.ts
@@ -42,6 +42,7 @@ export class AccessoryInfoComponent implements OnInit {
       'Television',
       'Valve',
       'RobotVacuum',
+      'WashingMachine',
     ],
     [
       'Switch',