+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - updated icon for door, window and window covering services
 - accessory info: show props on characteristic click
 - credits modal: added new section for translations
+- custom types: expose a switch/outlet as a washing machine
 
 ### Other Changes